+++ Evergreen/Public/Get-MicrosoftVisualStudio.ps1
@@ -26,17 +26,12 @@ Function Get-MicrosoftVisualStudio {
     Write-Verbose -Message $res.Name
 
     # Resolve the update feed from the initial URI
-    $ResolvedUrl = (Resolve-Uri -Uri $res.Get.Update.Uri).ResponseUri.AbsoluteUri
+    $ResolvedUrl = (Resolve-SystemNetWebRequest -Uri $res.Get.Update.Uri).ResponseUri.AbsoluteUri
 
     If ($ResolvedUrl) {
         try {
             # Get details from the update feed
-            $params = @{
-                Uri             = $ResolvedUrl
-                UseBasicParsing = $true
-                ErrorAction     = "SilentlyContinue"
-            }
-            $updateFeed = Invoke-RestMethod @params
+            $updateFeed = Invoke-RestMethodWrapper -Uri $ResolvedUrl
         }
         catch {
             Throw "Failed to resolve update feed: $ResolvedUrl."