+++ Evergreen/Public/Get-RingCentral.ps1
@@ -29,7 +29,7 @@ Function Get-RingCentral {
         Uri = $res.Get.Update.Uri
         Raw = $true
     }
-    $Content = Invoke-WebContent @params
+    $Content = Invoke-WebRequestWrapper @params
     try {
         $YmlVersion = [RegEx]::Match($Content, $res.Get.MatchYmlVersion).Captures.Groups[1].Value
         Write-Verbose -Message "$($MyInvocation.MyCommand): Found version: $YmlVersion."
@@ -44,7 +44,7 @@ Function Get-RingCentral {
         ForEach ($installer in $res.Get.Download[$platform].Keys) {
 
             # Follow the download link which will return a 301/302
-            $redirectUrl = (Resolve-Uri -Uri $res.Get.Download[$platform][$installer]).ResponseUri.AbsoluteUri
+            $redirectUrl = (Resolve-SystemNetWebRequest -Uri $res.Get.Download[$platform][$installer]).ResponseUri.AbsoluteUri
 
             # Match the URL without the text after the ?
             try {