+++ Evergreen/Public/Get-VastLimitsUberAgent.ps1
@@ -25,12 +25,12 @@ Function Get-VastLimitsUberAgent {
     $res = Get-FunctionResource -AppName ("$($MyInvocation.MyCommand)".Split("-"))[1]
     Write-Verbose -Message $res.Name
     
-    # Get latest version and download latest release via SourceForge API
+    # Get latest version and download latest release via API
     $iwcParams = @{
         Uri         = $res.Get.Update.Uri
         ContentType = $res.Get.Update.ContentType
     }
-    $Content = Invoke-WebContent @iwcParams
+    $Content = Invoke-WebRequestWrapper @iwcParams
     
     # Construct the output; Return the custom object to the pipeline
     If ($Null -ne $Content) {